# design/openapi.yaml
openapi: 3.1.0
info:
  title: IdeaBank
  description: IdeaBank
  version: 1.0.0
servers:
  - url: 'http://127.0.0.1:8080'
paths:
...
   /v1/api/idea/create:
     post:
       tags:
         - ideas
       summary: Создание идей
       description: Эксперт или пользователь может создать новую идею.
       operationId: createIdea
       requestBody:
         required: true
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/create-idea'
       security:
         - AccessToken: []
       responses:
         '201':
           description: Идей успешно создана
           content:
             application/json:
               idea_id:
                 type: integer
                 example: 12
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/idea/:
     get:
       tags:
         - ideas
       summary: Получение идей
       description: |
         При указании параметра id - выводит конкретную идею,
         без указания все идеи пользователя
       parameters:
         - in: query
           name: id
           required: false
           schema:
             type: integer
             description: ID запроса
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешное получение идеи
           content:
             application/json:
               schema:
                 type: array
                 items:
                  $ref: '#/components/schemas/idea'
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/idea/expert/tasks/in-work:
     get:
       tags:
         - ideas
       summary: Идеи в работе у эксперта
       description: |
            Эксперт может увидеть какие идеи 
            он взял в работу и не завершил 
            т.е. они в статусе in-work
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешное получение идеи
           content:
             application/json:
               schema:
                 type: array
                 items:
                  $ref: '#/components/schemas/idea'
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/idea/expert/tasks/unassigned:
     get:
       tags:
         - ideas
       summary: Не обработанные идеи
       description: | 
              Идеи которые еще не взял в работу
              ни один из экспертов.
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешное получение идеи
           content:
             application/json:
               schema:
                 type: array
                 items:
                  $ref: '#/components/schemas/idea'
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/idea/expert/{task_id}/assign:
     post:
       tags:
         - ideas
       summary: Закрепление эксперт за идеей
       description: Эксперты просмотрев список идей могут взять 1 из них
       parameters:
         - name: task_id
           in: path
           required: true
           schema:
             type: integer
       security:
         - AccessToken: []
       responses:
         '200':
           description: При успешном закреплении эксперта за идеей
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/idea-assign'
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/idea/expert/{task_id}/complete:
     post:
       tags:
         - ideas
       summary: Завершение работы над идеей
       description: | 
            Эксперты после рассмотрения идеи 
            и общения с автором выносят свой вердикт.
       requestBody:
         required: true
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/compelete-idea'
       parameters:
         - name: task_id
           in: path
           required: true
           schema:
             type: integer
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешное закрытие идеи
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   message:
                     type: string
                     example: idea closed
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: | 
                  Если учетная запись пользователя не активная
                  или работа над идеей не начата.
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/idea/notifications/fetch:
     post:
       tags:
         - ideas
       summary: Получение уведомлений
       description: Получение уведомление и пометка их как прочитанных
       security:
         - AccessToken: []
       responses:
         '200':
           description: Список уведомлений
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/notify-list'
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: Notifications not found
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'

   # БЛОК ЧАТОВ
...
components:
  schemas:
...
    create-idea:
      type: object
      properties:
        name:
          type: string
          description: Length must be from 1 to 50 symbols
          example: Cut trees, plz
        description:
          type: string
          description: Length must be from 1 to 900 symbols
          example: I need you to cut down the trees around the office.
    idea:
      type: object
      properties:
        id:
          type: integer
          example: 11
        name:
          type: string
          example: Ya.ru/ping - to long
        description:
          type: string
          example: |
            change ya.ru/ping -> ya.ru/pg
            add response body with 'pong'
        creator_id:
          type: integer
          example: 5
        status:
          type: string
          example: closed
        start_date:
          type: string
          example: 2025-04-04 20:00
        end_date:
          type: string
          example: 2025-04-05 20:00
          nullable: true
        expert_id:
          type: integer
          example: 15
        solution:
          type: string
          example: No.
          nullable: true
        solution_description:
          type: string
          example: Your idea - shit
          nullable: true
    compelete-idea:
      type: object
      properties:
        solution:
          type: string
          example: Declined.
        solution:
          type: string
          example: Declined.
        description:
          type: string
          example: Because your don't give me chocolate
...
  responses:
...
    forbidden:
      type: object
      properties:
        detail:
          type: string
          example: Forbidden
...
    bad-refresh:
      type: object
      properties:
        detail:
          type: string
          example: Invalid refresh token
...
    idea-assign:
      type: object
      properties:
        expert_id:
          type: integer
          example: 12
        assigned_to:
          type: string
          example: Change ya.ru/ping
    notify-list:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: Idea [#8] completed
        date:
          type: string
          example: 2025-04-05 20:00
...
  securitySchemes:
    AccessToken:
      type: apiKey
      in: cookie
      name: access_token

    RefreshToken:
      type: apiKey
      in: cookie
      name: refresh_token